LayoutTests/media/audio-data-url.html

<html>
    <head>
        <script src=video-test.js></script>

        <script>
        
            var data = "AAAAGGZ0eXAzZ3A1AAABADNncDUzZ3A0AAADVW1vb3YAAABsbXZoZAAAAADGfy3Hxn8txwAAAlgAAAbmAAEAAAEAAAAAAAAAAAAAAAABAAAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAIAAALhdHJhawAAAFx0a2hkAAAAAcZ/LcfGfy3HAAAAAQAAAAAAAAbmAAAAAAAAAAAAAAAAAQAAAAABAAAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAAAAAJGVkdHMAAAAcZWxzdAAAAAAAAAABAAAG5gAAAAAAAQAAAAACWW1kaWEAAAAgbWRoZAAAAADGfy3Hxn8txwAAPoAAALgAFccAAAAAADpoZGxyAAAAAAAAAABzb3VuAAAAAAAAAAAAAAAAQXBwbGUgU291bmQgTWVkaWEgSGFuZGxlcgAAAAH3bWluZgAAABBzbWhkAAAAAAAAAAAAAAAkZGluZgAAABxkcmVmAAAAAAAAAAEAAAAMdXJsIAAAAAEAAAG7c3RibAAAAGdzdHNkAAAAAAAAAAEAAABXbXA0YQAAAAAAAAABAAAAAAAAAAAAAgAQAAAAAD6AAAAAAAAzZXNkcwAAAAADgICAIgAAAASAgIAUQBUAGAAAAAAAAAB9AAWAgIACFBAGgICAAQIAAAAYc3R0cwAAAAAAAAABAAAALgAABAAAAAA0c3RzYwAAAAAAAAADAAAAAQAAAAcAAAABAAAAAgAAAAUAAAABAAAACQAAAAQAAAABAAAAzHN0c3oAAAAAAAAAAAAAAC4AAAAHAAAABwAAAAcAAAAHAAAABwAAAAcAAAAHAAAA8AAAAN4AAADjAAAAzQAAAOEAAADyAAAA5gAAAQQAAAERAAABDQAAAQEAAAESAAAA7AAAAREAAAFGAAABQwAAAtoAAAE9AAABUwAAATwAAAEmAAABTAAAASEAAAERAAABGQAAAScAAAGVAAABLAAAAU4AAAEiAAABFgAAAPQAAAEFAAABHQAAAQUAAAEAAAABAAAAAVwAAAEIAAAANHN0Y28AAAAAAAAACQAAA5UAAAPGAAAIJQAADR8AABJ1AAAaXgAAIBsAACZzAAArpAAAAAhmcmVlAAAACGZyZWUAACyLbWRhdAAAAAh3aWRlAAAAAG1kYXQgaBABoEAOIGgQAaBADiBoEAGgQA4gaBABoEAOIGgQAaBADiBoEAGgQA4gaCABoIAOIU/xPwfm5VaCr/AGCnoo4bXQD/+758zV6ePaaP/6vxrQg4wQbW5PIdnWflTClglTh+oZDB90qMHlXHdSxienzi30UjHaKt7a5tOW2uaqfO5hgyQYUmH9rgGsPSMcOBAwWws/B9h81rbF4hgAPXPC+v0nrzOHFlN+/3tz+4v6bch3QNPcaekXH+TYpv8T/H0/LFpg2XebZuCyGa8oACneCdl7SPtbOLa7S1CHnNQ61IGEOlU/r3ACBD8oqyEMCKgyYIPyz41snf/x1nS7pBtbk8h2dZ+HNeSNr7BhaeQUx6SU3GIpMJTFBlaNN924sAOAIXpVow0Gx4CqGChDIRhEAl6afBtAtw2d6YXLIm4AJeU/gWc3k9MzI2fPw3ZtwNMSzQ7N3fyvhkiQ4I4GBGe69VcPqRQ4sAzDj0WzcYmk5YlWczj6U1nHPFoVFQdiH54zI8KnGigUA/daLG0HaqmkzjokyUfS+PrjRdpxw+Mz6VQEtPWdsRgcx3/7+EJzkvYATsIyOGZWtQAT9yHydDQtzZyDLfyTbbGtxXYswZYP2wgIgWxmBojBRhARiAIhAR8F/XLitUA+u5YXHtCoXml3ZABeIN4nr+g7TDRkmADgIRpVXQ4GI2EQoEqmCiACZAELhwGt7DhgUg00aESIbaR3gSAZIVYgMDgnMGRu3xmR19X1imfyVNOaQj93yBRqf6/dyu4xAOc4AVm+3s6QAgecyrv4dcyNBLu+6MN9XDewCqrMYWT2V9TDAO8LhRQ8GD83ec79q3UlniQzdRxZVEjED23d8cYlWVT9NxLO6YCUh8w8WaqbvLOxViuiD68rjzYNEFYBXaibzRcNs+MLATAcQgUwgsxgEyAII+LtXStAsthCpkHfBMD22DZgKpg28jDIfLREKZYK+itp7QALIARtgDghGlWbEdYCdIBUgoATWvYXra7778uNsaGJE1TQACe/uBpVwRZX5aQ9zqaMGBKM5zu8Jg0imJUsoWqYrjfBzAnnAJi3pwuHltN75ASGOPi7QDMGWtPU/e/7f25m4mUY7vQBiCQhiO//Gy6zAKYrNZ+H8dAdZJU4K/jkVq7/6/pEAsBTp7Ll6RCXnGrLhqMwCoZRS/1CvNYDEIEcIMMoBEICJ8AM+uXEzyDdp4cmVL5YoBBrEKjxAzywgcYCwqEoEoT/7aIA05Xq4xmPXjuHISpVcIYKp6EAlOAXCI3IJgEdaHky1c6TMu+cnWTLa13xYZiv9nG8/wkGFgJvrGnp/cII3UkuEMQmozcl2khEogx2JtW3/rJAAItq3/50agAv7YCd+rgJgCAVv+H7oAjQ8X8KQB+48n5UJVbwcQ1cha4UVq/AyAFAC8S6dMAFeHUCKyY1+34cPz7OGyIA4ARJIFbrpjnw7haG8CGcBiEBCIAmIAiEBH3aTyKV487Qxd+d6Bdn9ERfxIIZfO4NKVgF9Ps8wellwFhcATiBUFOYWYgcU7VTg/8ykA0iXJDKCFQcIU//U8FTBXo0BT/H6d+dX3p51EBouIB6UHVqYIaDhUwBofNt2T3utqe9xHfEZ9z5ZyHyX4/z/i8nk5rbcXprzCEIQDK1NNoS7PizGRibKrphVcZq7vnFqcTrHUOJHMdw7x3E42kpBAgQtIlU3ei6WZLORZouGxCwZJ0ne0bWcOMjhmUIuaOBS8wLSk+XHlVGtz174HbU2hgi16UwilIvvx8/+8oJ2HFQXAANSg+4fHgVQkCRDg0KAv8+OtQJ3TuBsxm0+WQBMAAOE4Yx4YoHVnDFrxfuxShF1Qn1w7FY4COFx75L0rZ56T+QAAAAErjtAcAhT/9Vsk4IohJkCtZLqXJoFfncmDc+ZvWB0rVb0uBMaZQIkBgSe/jeQyOYDjM/FYQW3PejIfDkdmy5HWOGUNe/h0qvUptS3Do9DOd5h1ItjkJDqKTZ9hS86uGeOVCf/XojbLMECqqpIqz6b4zzQVWb63yqQzy03nyyU1uoZR9QPdboZ+2ysq8aaase3C++WG/X+MdePHRlHkri3KUpPZwOBN7LkjsYDFQdVsiMRCIRAKwCiKAxIISBu4ieQtVE69szlttOuYeHDSWMpgBrDHEbjroKhRKMAKuUz5zCAAAACIAqALAAHCF6VXQVjKlCKgRkIRAI6DWA3axAQotpAc7CeC1cKteomEkEIA/6epuE3FiOnQ7vgCM4vTcUmGY6MZd8wv4Y3+gpSJZitSMOPuMd9YwR9GIsmWAO+rpiiGKFCT6m1ZOL4Eu8Z/gmZTIoWMX9OU5QSEZgVGslJ3hRJBXVqZAdYWVwm6ONJptkTogIf33Grodt/e+I+jpP0I88hzJdMjvlDslW3sdTP5/EcnGJyCvc6FNYCEIDEICE4CLaaUI2XgjijU1QADEEPab+EQlxIMow1nwLiaNbP6HDt51+augGvwWFYpAQXCwvyN5YzgU9QCYBEoqtFNcaqcsail07ev+EoYcguXAOIRpVa9EKwUKohGowEry620wMRXGIvCwXnQDkZbvRs/M6qqwyvD3HpeyrwIqmZjXInC9W2XI+J3nsylJVxpayWpqeJ2vqs4CTcYQnPPT3eiAT1OWWnSgSPkan3MG69Uenw8ti6N39APAzi6NtMDnWd0EBgfd+1ROq8Vtf7wO1t0zv+Rv7fHq/TFdM0bDsCm8R2+4VpGTV2nzt5/oEqY3VvxYZ0oADHdnNLqBsnPr0RbMxYKtKUSjdwXm1aADjeACrdAyAYhARnARpcg7CwpEaqXLACQ1jcVxt71+FTKs0Xnr7+mlXsEwuE/0iIjFugA3fWAAkHFa9B+RhIgNKwHSWgBRcLicVBoa1FwyyBNmAl3Q4IRpVVRlHAiFAVZAVCgzIgjEAknw5Eqcs6EwpXRQXxdBFEqcX+KPTyjmKIDdCaHgqv8vzHphCOMKZr+UjDO+T3hyWj55WgZNh2k6SybXdImzPH2a6MTvOTxCkVnh/jxiqioVgNxiP29uddrOYAEdfTWsVnOYqDjA8ibuwtMgKFlnNLqj79Hz0MAUa9aL5vQ512aACOlWMwTIqyk9Xn+i4NdutFbTjPVYABMs37EoSAlHW985f0LTipvlSJR7nfolZNnG27PjRDvArTuBhMATGAh9Pd0EAVKiGugAr2fD4XN+UDYGbLQItDE3lA9SoA1t3EDNiwAFsYAFwFAFwFQEu6WNXQaAG3KARmgB3Q4AhGlU9jYVhQcFWYmARLEYa5DgdtWpeWS5AA2vTkJ1AqN04ESDLRfJCd6Ufvg6WpXh/lyowgN3buZtHDOoDnDGEJ5HwduwcbsflIBiqxcIK6BuVPv+rniNzmlRw2iodfx+7cVICAcertGxBevCYn7f5a7MWnRUrvFTj/z8Z0KsKabzm7LxKoBDeM9tigqM1GpcFe2VXE41ZRhyU4LgIBi+X5uv4CnmemFxXF7fPP2QZ2vUlU5I06WD0gq1WawOJACZQEOr26zMBZZeJd2AEhS7krXV5flIC8gC+kBom5gHWhEeY2aJpmXPYir7EQAAVvANABsRA3w2wZfrrVrQACYBEOCEaVVUWxoKBLJBmMhGEBHSw5QIM67ssWNSwRQeA2uu3KZjYGm/UtacRXkVRU0y7d8HRS2S2UhXU8+NGmsr4TcRkVel4BLjQrEM5+OtmbEPRjqjvpZuCui9u/hrr2ApiPT1dXHp6AK4fPs5QF7kqo55+O7LwXfGNt3SiC5zz7o7KgBjuF/R0SCoGXfs9mD9w5IJ25SqcAq9vs4YOc6/gwrGbRVRvoqQ45uo1vPL8m5eFHtUMXzIeXx143D6YhfX5ZT9DQhTZmVh94VFK91gM0gIQgUzAEQgIt7GDbRrLZxE00AsGIGTHf3kFZirAQrYuDq6ANTtqgZvo/ABYDUCLW8zlBO4CwDkmuG8qz1A2RIBg9gchGlV0FlbBdopASTzIoVV6DbQ3wbXay7DjLQt8x2AT9JCuOu5viGvzn5Lo1d62Rj0WQqtsOd0uK9D/Q6YBSWWl13CtNF55092eQ8IEhyEXe/0RaQu146G7fhpZAlGOGlkEUYUxji/IgC2bHHHk5QIEY2utfQgBAnHwsAoJfW4PdsvNGWaP5sC2KC3rFkoAU/QXF0FACQE3IJKIXD+IKtWGsBCECigBCzyOxbTBk6NI0Av1FRkLCvbKWRPELvcqwKwASgHx6GEDR3uqJk+vgJQAJTBafYFswD0iRLn+gG0lyOA8m0ABJr7xuKIhwCEaVTwgRWIAsRWmkBEkp14lr3c2LgXKaZImg5wtB83ci+AiDJ5F86siQZtvsvBMADgnAhK2OZLWOxNZggil+qVBDuFSVcWQEIAC6vqonojw6S4CER9eM8d2C8xiFfwj4e6tIbzUsRygEqKce79+lWXN1AnRWBNl9uPn7SAXlfx+EXCox0k953Z3I/SMXNfGp5MNP+DusDdsPSXylwKLhveLhXzqMkDlFRTKAYrjRalOGQhwQCuVZsARuAR0u3A8F2YJZUai0aAVYzgFL18PpTQmt66dlVete/0UkEzYDU/6H1CoGaumMxk6D2AcLdkACYFwWAAoBq9QVRNCxEZlCI44Kj3qE+Moa/J0aAIAVAO6HCEaVT0aDORlKRCEUBEWXBhLCZcpoxZLaER6Fs4bLyoJYaVasA4AggENsxdR+s4FOYxgmVfeLMhCzObR1NYoxlxdiBBrvz1M8K5GkwY0XfZXj9HdzEyQBZTMLqN8mCjSAoAzXKTno93wVAgAcggrAJj8ZahZUQ3aeEFHU9SRSoqTNxpee7t8ZvU1AVc33b65hpHHCpQ+kntD6j2nkiKa2uF2y1bVV6Vymgj2HDK9haB1qn58E89+ac7UCzM+jaJ1FFgljDlcKjja6JmgCYz0vqWLsR5KAAqzSwjUojcAVCAi2J8I2ELq8NFgW8g9qvEdj7YS6bwyPk0FXpG2jjOd0mAGbt+VUC+b7qCriVeX1zMLAZAzNGUNLBIA2IBfN5DCeZglYuEseJcevxRTNrQ+V24gASjgiSo7+BPiBvKZitslWAuPPIA4IRpVRBJEg4Eo0GwUMwkEwkIIkEIQESM9kMSRe5EUOMSi9AcqR1vfzllcRjUor6mo1v9jCvAw2xs/pljkYrspb6P4Dh78yVFx7v8b5MUV0UKJtm47Af2w7WA/XxfhduoabeGqrQncUdA1tFObs33VIeElN+cpPY4a1qkJKyaJn+VBz0FgTxyUbUaDJlNBAt5FMxCcsw3pIZ09A/dquUe5vVMQchzzF+oqae5Vw1xRfNqmTXmyRjWWjdktO8LsxMVgl8Rw3WILdzFLvoxAoKXa/gSfD83mJ080QqpWaFGJAEawEOrGmJu73pYdvIZqSuADIBMGXzWyuthIKUwzVlE3wT/hJcfHAzCtbkFFV4jW7gBUSqfOMtvsaJqALNOguno9cbmtl0I0CmEGt2BQEQNEakRwFIn5JO5mMmrUABUljCeaAcAhGlUtrpsGocGAQsnnPHQWTEtzC60UluAAKdLQR0X4+3dT+/pLUJMpNdZHoyO/9/qGXYwga1TCGJ4zLOHk2H9Y5rfMCj1T05uvukf67kS7BZa8bpMwwv2spxUJJ5Y01ei9E2iyPj3H6Zxg7OqfbFB4ChthrvmTNco8AGxhLhSYXX6Nhx+3PcbfrbIrM1asNysoqm59GlPbOCpGcq1+Fpe8+Z4zzyFgvj6aL2jdWKm4TKwmo19xu/Ho9DXdCfNjIP1CaGrS0/PoCggtIt2USTVjbWRMubiK15HMy7eW1Jp7NCa+8/Nauc/a3W8FWeQh9odQnJ1i51xHEM4RRZ1RmAt+Eeyh5/TFj47LS7erb2SXSKre2RUGyJat59ADCFOXt2GcimUpKXAmodCfu6qadl9gqJLJuLlgxCU2/k0LyrfEbqIX6gkqtnjQITPFqC6hKgPjqBo3Xw85mlkyJ9GPxZ5M4C9HNj79dTKvoiu4w7iJTEpEpNqXRn8G2yNpw5+Nm3WxlSaincRxTCacq0Oi26rQlJhLIgbFgICWcZ01g30lsMiLVd7Jc0tpP3V+BzuMB7ObXq4l/pdwdhIxMJptMlVNa9wPI42rElt2+mvD+tZGb4kpKrTlNCCjDG5Aw56m04zM8NCXSZkaeOO9j7TWyLDu6HS722qSsuJOo66fMSycaKHdsXWlpEsWc+JPLNJ5sJksk+V1BXpTPwnTQc/LTxn14eEuMlTud1+WdsSvJj2Wd/Nq+q7Gl9dt5Ltqq6rqbptG1meh8dXmq0aACsSm8KbGrF1FXaaIFnJ11zqKvWY9THy8SftZyLVnKLrOOalLgGCOMh8mk4gpNs2pLdkSpZcs9RfO9GMpZ8/AeZX3BA656Vw2LfnJMD01Gl578T6bPwuBZVT2zfPw8m22WF8OGjunLVxw8PPZvksc7s0HPKXqrp7loY78NczULk1U2yRkOiR4iJDDFPASIbj8IRpVRBLGqEUw1GQ0CRAEM87mwpWiC6QIS5SByxhj5oVB5nqBD2JTlFV+Ncp2qDp4ZWQmzHdFXxyi8dnvvUrrnngIZWv3/79Jla3WlBaS6PHTQibHnmbniuoI4L1n4yhdQ4ce/KmMaRwY2bJIv3/ny98gsCAQVi06sjHMczm6irIqO58K9FygRWYzEqiXh8+zr3mDgSvPly9/MMiDuS879nvDCFHFEFhtqirt3ioHwvh4Ze8Nm+Ij2pGWwH9/u0lK321KlBoQMBm3Th34WM4dAqnKhzaBxOAgiWNqOADgLoAgNxGEo+jo91fOk9D/ZdtHn/sWq0bIUlozqgOVaX4f4v1nTZLy4dDrrRnbX5UGdOZtsulLBj54fsTmw/j9F6XacZlUnV/ZFEq31hJVOWO10y+7DsyPfwv6pfSoADghGlUihorwgSBoIBUKCUKIARi8GWxu+oU6YWyyyAT7e848AAxi93UuMDMPC1b7ujTfCki3tObk6uyayuKmMOp0lxVY2UZbGLHRx9q0GFzCKwSzwvW65jcqiKmMsZVKtbiVlFsccVQRzCCi88MUHKQWTHHU10tEBEk6YjQdPXYmi1dn2bnJHfxMw8a1011BdfI0Jgme8x+UVF6Jldy24fJ4g8Mhop5tjQ6pfQ54vLSvvTeSmYdOuGYLdyb1UYVWoglUJnJsJTr4pceyW7mK2d5JmtUGR0k4ONVlVd6aToc5CYBhrs8tl9ugRsvSq0WAJzIc/FvhHiv8f0+fMEg01ywv820R21wbhbRyeiDorBG1J6v+L/7CDS2pdsvsfRuvT0pR4qulagARAGQEC8YY8Wagg3gBrkV8Iw0qwJqiYJJ+T6a89F8U1zbwjr3cZxce0uC0xc4hGlUYho7Dc0BYclALnIoCTu9UhKc74dAhVgu4Bbqm9HoD1CyFF4OMOLTfHeZdy9JBNHrkRCXJRSiriyAfVJELN6Kqb4cGdZ+rgJJFAAsNr9INkqozMdCNeAMjh4T6m28SxhWMQYVj1WyauIu4rWyVGa66D6J+/aVWqrrDqN8IrNdRE544pjSy1M6kEsUWrK89/h+y2FyvGGcMWNUnUkibwWpiVnJloixdDTpgAWg79v3S/5g3qdMIm5upVXZ7/B4717s55vp/+8onVIQyR9eCEbjmAFK6xMwRQAxUAgvay8OT48PIb4YmrIAuqqXUM9W07v93qnBJkyq8ogqx1uNKKUYeLSAPXz9b9wEY66fbLq/meerLbQgpyHrjOsxeLETzgAVAJgKxE7IZwvLTuheJhzp46iBcDKLyIKA4IRpVHRmM62DI1IYkKIQEDHF5TG71EHGBFTiSIIgAC9atJBwWR2N2hgdBLDMfRQ2xCni24iCBRBEd1PJqCiacBG8q4GrUl449j+N1gCmOQIcT/NMXRbKcWIInsNfbFIorIJI2r9t2dzHGLqRTDNq/cvavObbm4ilIwQmfu1tYwiJkUp9fvx0bGJCZROu3x/l8tAJjzzGbZea0B1r8VCpxNjt3pAq1TpVJKRRnizcRy03NQcjjeRysEszsa3aFoXe+vO7mknvtjXd4U1ApVeBTMAxCAjOAhRPYzFZcLVxckkAXMor7yoXX84qS8kZ8Janv9/UzXCM2BZzkCaASUAASE//TEC1Rr8AMr+tkYRYIMy1sF5G/4DOmyw0dDToXQUkklYIBmgHAIRpVJZIGxSFBmOokGQkEQwEkzueSl0FiCFLZqy4tAnUONT6UyYDqz6XiX3iedfyVAFOaoqYCKF9jvPqew0y4KmWdEUvBQBc/Ua3dRQ13bt1Bme9rsCAqBDjucCEtyAXNikVGxOdAwxgPLnlmnJFIwvykcIiZ4AESEtnBF39WmDhII1lJCZwIi3VmkQSKEThSrzVo+xOs7hYJ8REKnlhoYwKcxc5X+ODCLFpS2mij7I/iQAfbEfLhpCxUs8By6az4jGXIiO7rqkjKvXsYoraRhADvYfzs2rOr9sGCrkn8PXi8s9oAAVUKwysERAMUAIQkIRgI3o6Wdh5blXas4OGaAOrVSuIuNCI3IxoT14zZ2M1f0ZLgAygPpnVKiyiOGuZFVGg3fz0AWhS4D/n+ceuErgdBngA2EBLuT1ndX5ExPksei6qgAbLiIXEASBwhGlVtHQdBgiqgKnMwCXbiGrYwvduZ0QLwWvQidkDx/hlxIr1RER8UEsCOPet6QDQHbD7P7EOP8qRKo//pwoSdBQVBnoPEpFiMBysnQkqdqW4mJp8uVZP5cJRCpnf6qySF5mW4HVvdHpiJt8vPceXBtnTBwU1qCjpC+H0L3mZxtG6+2CL/VuIQKi7l2TAVsk4omjZFR+mrmuCkzvHY58A9nK4v0xBAVrOkKEVwlSNdiSuTVee55dzayxbtfUvmOx10OTehYdl2zmLKWkFeqzcoQGJwCIgEX8EWO4WYaaadtJdgLgRSc3Jt78Fx2SRBSxny/zlkCxv4AdA6IiLgFAyFXv4gtVY9OtbJ/aDL4p4VI02gAUCkC6wAlMAiAASACQOAIRpVQ4aM6IGqUEpSEIwEaI8dWoXeClqXEVLVrQJ+NbdgACGbcpdxHoNoUkQk6zhnRRPsHkZwu0xlfd9HmMN7TxYTHL58CWXovyKEjKCqAxhpsCA+4HgyYMRvXu6e3qu65BMe2XmnMpE8wnP9od/ZWYnLFcuOR547/YDW9MPO6kah+XUVVmcRnMptvu01bVpGeVSf67z33xVp1dNzffmh/tpMROJXIAsMhMAg+Ld8gh5A5X04x6Ovh22xShjDJVhrtpcRKQw7JZHjFgqFeawGJwKLGOhydUw0Lc6kAO/7Co54FaXXH5C5kAuB5fYGhc1rKQZAH0/652YRoBoBuzlphajWaVCq5SDuAGAAJgCYAChwIRpVJY5KQrC6FcZgEVi04PFai630JAgWXoAaoqlOUCV8IRD+6kA6PmKa/1raVu1elXh1WZ4jWHgLhgyvEhGPJ2WNfsvH5i4FQjPq/R7tqfla0lG74/x45d1hql12eLsxiRH3Cut7+y+/gt39JV/OUOjlrcbhzw+AfWKmM9tpTjAOM1quOCOrsmsxgPXLB+iozx+tUO+3J8zB/OottiSbShO1KVSJbrh5DIpm97xNj2+wprbFPZWm2mcy2Yk1KykDY1JlCQKh1mxRioAiIBAsfXhgHAtg4yk06AVu60TuAFdvIQAasZ5MO8gMSZ2QDh91EbzeVQKPnrjMHpyhzBIAiQGUAKMUVy6UHtMADCAKAFACkRAWwgAFAAchKlU1GdbDcqpEoCRAGnM0WWvrsRlhqwAARGe9f127MNDzVmbNj7OiGWb9DYgEIa+ds+NhAqIw7D83GTG7T13K0OVuxKYepyFQXeA4PXXoD+2EQMhQvj7Tra9LnkbLxX7vgxsm9gx4NzefPshXW9tjlc+xjJHWW/tPf279Py0Xga/nusT3fciomaOsRqVOvDeujs60V3C/7ReddWVz3fco64J64vJ+vGG/LVzdTiVcmKrEigDIhgSPycjyvnXFplj8YZYhBSLUBX8AAFYalgZwEIQEIQEWHwKQQRxbtqMX0Dzsxh58I3Ft/p+uwF1uJjhSqrtAANXw3uu2xY7dMiAoh7iSQwXrw6aO+C6xGYUrV6aAkCCR8dF9OFaO/B6xFIGMADAWFAAcIU/zUkNpBYoxQGrGORy1MgV9WV7Dh1/fXryX1RfWn/gLhbVkDZ8fN59uLUt6q+2SyekamtR64II4ZwaxCK7369pm+Vm8q5B0hU0iVDsKJPB3ix2KhkZDAy5ZUI1EQaCC8paGuigzBLceR+maMiefTU9qNdPYRhAOpDhwbw4WI8fIkOV/vQAqxUz2iFNAt/IBv2duGif+PhFTLqEMDJqaQmGMNsRYPeXKVyHfr3cA4M+eEBI8DUjMRlzX4K9yg9M5+n8r5HRcuoBwNZGziG9b+omUIluOwXCZd4nLNeamS57b8FVnnKXf996cVFu1V7LLEtQ+rfqlt7swAAGW/qwzadAhLml/fGnL+HkDo8R6T+DFIrKMRwFkXAZWfgE8zLfj1bief+OvfrRxv9rNFGpLQNnx82ibUidsrbQcLy5Vbww8VFuE000003sRApgOO9invq5y5kURC65zyErTUKzgO1Sh3pDdBCDJIiAGJBQAoEgoEQ3T4GEpjN/0MExqqifR2IY/vp9foYnZEwD7fkzEEw4YANsOIXpVVRoKxYCiWCjAEgChNGwmWRENMuxEC2Red4GmEizJov8XLbpO43WdKvbFbNIaEHJrGjAHOO85ZexOYqxBmHVc7jc/c4Ulii3IExGp0hZeg7w9fHg8J+iUvhQWO6a/tDQ2qdE70dCjsRSf1Bc1M0a5UFnc2TK3pvrgnFVYysFZfVT/LZpwV1MjSTRMZS5R1OFHwqWN0VwSqJM4BPcd4l1cDr2HXthNfBeWg4ys740k4K2KpinhJ0OnrEgoFyqpBtSi9N9hvQ5orztxUgq2QUWrml1YCGOarHeuEjBRBFAYlA4mAQhCA4xiGuFIDIYXAl3DjhnVq5Clv3u5BR+q56ekFJnA0wkbOzgww6gXtcBS4UCmHKuwMVO8uxjCPeRPbG9GWaYUWO11EQAcIRpVO4aMx1MwTLAUGwUEQ0EIQELqhSwVEQWmaUuRgPXxoOMgspSECKDOvlsGZ9uvqO9BnOjWpxPGGJpeiGhRxxNAKCk62aVVQHcpEXdLYue+amvd/jiPHLEtU58+zNG0kG0277tLTv1vDNtbX0LXwQk6VmmEaGujvGS3bBAEpSYSIbavwVhfADs23juhQdVrr11a5ow6Ipg/hHKqznq7NOeD3zJwqiXBN0Yc4RWo53zdq7U/GxAxZscMnbUIujcu9XC+ef4G8doYlaAIlq3aMJLEybmHy/ZMK2VoMUEIRAQUgI6t3M9gUGNabaUvSYAQHFvC9kY8v8aq9oTBixQ0tjK+TqQOq3xepDkTN4okY934ZPApREKE7ZFNJ79Vs94FoBgtbinzx3+AgAmCBUAeSlIo0rr11Zo38Np+A5aqI2jPN/DLzlDwzy3J6opAHCEaVVOKjM0hKsigJnViRtNADXKlxea0LBoaVfTmZQKsqJWDBLMDrgfEz0SMXs7+Wi7ASo6Aps2GHaAgAO8ebhIDGxGqrkkSAwAwV1Hvxl+18itiaOExKr9x8BAkjbXK0q/G3DltRq3drd+zq+fzuoEkQuMVCe/f9fhj59uMCQkCXHfo774ZilMhhMsbXPHGGs5XSoXWQrBvmi7zmKCVWBVlzJI58e/ha+bV+tWaxSNj/j2q+Pj25ndKBP3rQ8rznHIUqzTI8ESiepqL12LBr0AVirFwJEgDHtydXgseB5jfnWQA0GT78Hl+T59gSiJvXbKxPs7blfhif4xfG3L/EmW/H+fbUBYHVh70gqpFLAJlxdCVq6v55e9qcZIATd6CAC1OIRpVRCEE51KgmFAVGgxEQjEAjR3PIxl3KzW3BbbgSLlhzqp7V8o+G8H4CExLn8QO88lmOc6TgLOmDISGJG+V6XoFQzY+PxOs4mGFDf7zIVY3L2Xj3G9737dDZvh/DtSwM/D+njpqt7QTnxNifRCrM/LNK7GIGAgZPWhGMCa9SpTv1xJ1dSMZFhlvF74/773x5xna+ELahXnAEmDf4Ll8PDsUzZiX1r+i5tQtpUN8bLLNmmGrDUe/MxtZoTEceqQUtb7CcVB+81hNF2U3DHLo7cboWUKo9gIQgIRAEQgERAIacujEQIVZJpd2DfDW25/g/MCfSgy4A7TIMkQOq00BbKgZwARAJjCTAJXroFAABOkwsAAkAHAhGlU9Dg9BY0BUiGEKDIiCEICBY6c4liC+WsOC7W0IosHi871mcjMYCnm1YXM3hDcP+Iww2ep8JBSJJBSnACmm6efOBFM0vsAQaLtJE7BPhUIhF06IEQ1Bb7XdUCK5eIqBu1i4iPn3cLrf3+iYTWnwo/y/QnfGUwJnyE/ZUQLg+gaTL56m6f4OAVXLWgzRVmHlXdsCXB0zMetDtf5Du6BSHbpPn1qDXJgz2dkYAm0EkWWC4rS0TLYnsFMtVOYIE1h7YBFInne50AiHn7HwNO6+dwtr/wtmFWqwMJASIQUOMBxhqwFyA3AL9oaCgB1e1gABQSHAIRpVXRYNI0G5FNAWERUCwUEAi+MSBhcAEgtLiWIoLObTtXBa3DcTOxpv46TvkImOEWgU5a4SdLFeUWsIRHWyTKjzXSLjHGle3+Y2IRdVVBN+poaT4a1+6ZIi8aIBgZbdKK4Xjr+/Qd3ZjE5tnG4WBVeXDOU4wzjr7MRz2YMMKRS9ja789xwY2Qge+h6QMxA5icr5+fRev3Jeb35Qynyfsz2VtdnMY+Y3u+zX2k4xsJIAe9Zg1urMtNwVCS+2uVPShzCTg4I5syFJSldTXouD5roU0bCqzRWDOjkyBreEzjVU7vXG4BiEBiEJh5NIrk0DJj5yjLyNWLFDpZnn53sgnYT/XIOAIRpVXQ4KxJcgZKpHGAi9EvVnucG+mVdoy1yaRYiCNWISX4YTdoUKXk5hjnxuhAyZQpt5P0AmUkPyH0KKYabpgcHpC+L/1dMZTlSE9lhXGKpjrsdvj/FUAYQgVnHL403tL3+XkATNwTrRRM43y/Xi4geNd5ANFkWq+Vl/iNCxAYYoiLsgvPj5/Z+smZWKldJjDHG0XtNfgYxWdYpclRYhmMTuPPxuwAJCVk4fncbRSlMYCwFeN7PgXisleOAIqc8MKlxfxduNRVUAVABd9PCgAAFaKUGJWCbAGIgCIwEOAfg99uGsaBMPN2sDDkfY2Ylga9q9F5WUccIAlHADY2+XDEBN5Dp2G1wCgXCaEwZAAFw5gAIKRAAlaMgWBpAHIRpVPRWRJIDKgDYXEZHGAlLNNPOetukxaiaTvTrVxYiBSpKlI+4HoSIbtwdpAEQO73VmkIvkq+XkEGECuiRAu0yWygsGbiQJr/3XABMUXPSavFSmRWDBOj/dxVYIFXVulwmOFlln4v8akZAAQAvV8xhUgvO6irhJWOys+l1v/n8qAAKACwMAoFOUycqQABQIhbzQoAJw8cBQXbvJi4tgiWFcn5dgEMcQVjccjHOceojGcdGroAAF6vydoAAKlSixgiQCmMBiYBAD7tfO3lexZ7atAEK/KUgHVIBaOmH5w+ZTCC2HMAff+e/HMAR6lQ83AADvhEAGWgCggf+iAJ0loVABl0AHIRpVUoaM4mFJqDLlOAlnBT4O+XUmGNC93TzNLWB7cHQdVODngogZlVOiR6SrGmgCDBEoertZanWSXJY+M019LMwX/I75Cgw5+jHBpNykA4rDV3tkQsXcMQxqau8iOh+2dJABC888yCQr0vHMC1QBndYQymP1H0jzGEC8wVcpmisML/NfomecmGGGEAASKxib1De9yXrQlVbQvHDWgAVMhW91Djj5aXRAAUWGPs/pqG5ne2hVs7nBcH652mMRCAAhIr3/onQAArjcA2CZAKZAKIwEx9A9juuhkH4WCVTCeTkTxCTJAGbyYQV/yceQAauaJkDwQAE8ADbEATUAAKRAOCEqVS2KDSmAqhhKkxAISxnTATw4JjgHE4agCCPTe2za3KSwz3IAi2zLy6eRrrBu+sCTt9BoCi93Bm4xjwevjja9TaVlv/3v69Oczr8X6nlMXUKorGgIFzepscWABIa6qblx6vpKThoOn4xESuAn9/WEY5TsiR5JiEiBO3B5SKz33dovv7t0e/XbNIxW4qNe7p/+5yR8PpxJeJ1ArQnbl0RrdXyKhjjUk306Kc+nOYAGdJF0DPv/z2ZK7LmRUAWQrsxeMSru7MjEoEsWXwf/FgaIAiAVIBYImAhiAZlAohAYhAohAYGraUrPP4Cr4KCoQGf4Jvdj+y9FAA1QoAREwBwhT39OQ26F0WLlKxSCQJx3HDR8On/9nfnLy7aaRbSoh8B57pCcMwUvXnLGmjrjBdsrVLaRmy9OZORzxep/iIACbYadN2gZWZgU0gA6g8R0FLDiiCB7XPO/RcvOKzxh7cR5lXKL+f1phzPN77ZZrYz9tQwqdmjizH/p5oHiUE08AGOJFg2gSQSEqoEBNYQBj95t/5MMIyS0pDc9Z12fszfzr074AKbJc0vp8ICnjZFaN+z6UAs6F3TUX0L63FSx5rKvE3XeFx+Hzr+90IooamUBC8n6+2O+OQrBb1alaB3C422GyI/OIwBOsh+GEwI4CpQ6tfWun/9y+M0Z1Ls6XgB+0qIfAZChAjFjaSA0ZmgAK8AOmtog+FvRKmVbkiDv4Fk15gvqYUXOAGKMgsAAVJ0quBZS0HqvDbFRG6nUr5sMar7Qm44FQavXOYxZr9t/xhgxy0KwvuAAGYIzAAchelU8GZCDBBCEKEEKGARvinBY3u3/eOKTNWXS2lzMnePgHiWFVKqnHC4Ql5Pc9zualkyVVVTjufX5vBHGnzeEyO4nTjiwO4gDY+bGHc8cWcU0CmgUl/NNZnT/AP8V/GYFP+UVZOjSTQKiOgqoTQVCaOjUX/KSsvmKhBYB5MINNCoUX6007uKCgo7C6zQo67wsuaFFQVEdBXG/FVQmnhBQV3sCqZoU050bQa/lWTo2y0FiCgo6513qhWRgaIwEQwGA6LMH/rjpFtLmY+AfEPEFP4x/h/ADyfxCnx8NjI37+4yZgAMnwDzYP4xxfxAB5oDPCA/mAAGTMAPeHJmAAz+/vD8Nn+DaA7w=";

            function canplaythrough()
            {
                testExpected("mediaElement.duration.toFixed(2)", 2.94);
                endTest();
            }

            function start()
            {
                mediaElement = new Audio();

                waitForEvent("error");
                waitForEvent("loadstart");
                waitForEvent("waiting");
                waitForEvent("ratechange");
                waitForEvent("durationchange");
                waitForEvent("loadedmetadata");
                waitForEvent("loadeddata");
                waitForEvent("canplay");
                waitForEvent("canplaythrough", canplaythrough);

                mediaElement.src = 'data:audio/3gpp;base64,' + data;
                mediaElement.load();
            }

        </script>
    </head>

    <body onload="start()">
        <p>Test that audio element can use a data: url</p>
    </body>
</html>
